+++ README.md
@@ -78,7 +78,7 @@
     - [x] `@DisplayName` 변경
 - [x] 코드 리포맷 세팅
 - [x] 출력화면에 맞게 OutputView 메서드 로직 수정
-- [ ] 패키지에 위치한 클래스들 하위에 패키지를 더 나누기
+- [x] 패키지에 위치한 클래스들 하위에 패키지를 더 나누기
 - [ ] Table 클래스 이름 변경, 클래스 역할 고민
 - [ ] 
 - [ ] 
@@ -96,13 +96,13 @@
 - [ ] 불변화
 - [ ] final class 로 상속안함 표시
 - [ ] players test 추가
-- [x] statistic 로직 수정
-- [x] Point 로직 수정
+- [x] resultStatistic 로직 수정
+- [x] Point 생성자 인자를 Cards로 변경 및 메소드 Cards로 이동
 - [ ] 
 - [ ] 
 
 - [x] final 적용
-- [x] statistic indent 줄이기 & test코드 수정
+- [x] resultStatistic indent 줄이기 & test코드 수정
 - [x] 상수화
 - [x] 테이블 생성
 - [x] 딜러 승패 맵 생성
