+++ src/main/java/blackjack/domain/card/group/Cards.java
@@ -1,5 +1,6 @@
-package blackjack.domain.card;
+package blackjack.domain.card.group;
 
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.List;
