+++ src/main/java/blackjack/domain/human/Dealer.java
@@ -1,7 +1,7 @@
 package blackjack.domain.human;
 
-import blackjack.domain.Point;
-import blackjack.domain.card.Cards;
+import blackjack.domain.result.Point;
+import blackjack.domain.card.group.Cards;
 
 public final class Dealer extends Human {
     private static final int HIT_STANDARD_NUMBER = 16;
@@ -18,9 +18,4 @@ public final class Dealer extends Human {
     public boolean isAbleToHit() {
         return Point.fromCards(cards).get() <= HIT_STANDARD_NUMBER;
     }
-    
-//    @Override
-//    public boolean isAbleToHit() {
-//        return cards.getPoint() <= HIT_STANDARD_NUMBER;
-//    }
 }
