+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.element.Denomination;
 import blackjack.domain.card.element.Suit;
+import blackjack.domain.card.group.Cards;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
