+++ src/test/java/blackjack/domain/human/PlayersTest.java
@@ -2,7 +2,8 @@ package blackjack.domain.human;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.card.CardDeck;
+import blackjack.domain.card.group.CardDeck;
+import blackjack.domain.human.group.Players;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
